Update example to go directly to MongoDB for querying users and roles.
The MongoDB identity provider doesn't implement the IQueryableRoleStore and IQueryableUserStore interfaces because there's no LINQ support in the current 2.0 MongoDB driver. That said, there's really no need for this as we can go directly to MongoDB. The identity provider is an adapter to plugin MongoDB, we don't have to use the UserManager and RoleManager abstractions in our application when they don't provide value. And honestly the value to me is just to plugin to the Identity framework and to simplify some tasks, but a wrapper around exposing an IQueryable interface is a useless abstraction in my opinion, even when the MongoDB driver supports LINQ, I'd prefer to go directly to the driver and not have another layer of indirection.